requirements.txt
discord==2.2.3
python-dotenv